# perf(Rules): RHICOMPL-3511 expose all rules to all users