--- a/test/policies/rule_policy_test.rb
+++ b/test/policies/rule_policy_test.rb
@@ -5,50 +5,15 @@ require 'test_helper'
 class RulePolicyTest < ActiveSupport::TestCase
   setup do
     @user = FactoryBot.create(:user)
-
-    # FactoryBot.create(:canonical_profile, :with_rules)
   end
 
-  test 'disallows rules not in the current_user account' do
+  test 'all rules are accessible' do
     assert_empty Pundit.policy_scope(@user, Rule)
 
-    profile = FactoryBot.create(
-      :profile,
-      :with_rules,
-      account: FactoryBot.create(:account)
-    )
+    benchmark = FactoryBot.create(:benchmark, :with_rules)
 
-    profile.rules = profile.parent_profile.rules
-    profile.parent_profile.rules.delete_all
-
-    assert_empty Pundit.policy_scope(@user, Rule)
-    assert_raises(Pundit::NotAuthorizedError) do
-      Pundit.authorize(@user, profile.rules.sample, :index?)
-    end
-    assert_raises(Pundit::NotAuthorizedError) do
-      Pundit.authorize(@user, profile.rules.sample, :show?)
-    end
-  end
-
-  test 'allows rules in the current_user account' do
-    assert_empty Pundit.policy_scope(@user, Rule)
-
-    profile = FactoryBot.create(:profile, :with_rules, account: @user.account)
-    host = FactoryBot.create(:host, org_id: @user.account.org_id)
-    FactoryBot.create(:test_result, host: host, profile: profile)
-
-    assert_includes Pundit.policy_scope(@user, Rule), profile.rules.sample
-    assert Pundit.authorize(@user, profile.rules.sample, :index?)
-    assert Pundit.authorize(@user, profile.rules.sample, :show?)
-  end
-
-  test 'allows rules from canonical profiles' do
-    assert_empty Pundit.policy_scope(@user, Rule)
-
-    profile = FactoryBot.create(:canonical_profile, :with_rules, rule_count: 1)
-
-    assert_includes Pundit.policy_scope(@user, Rule), profile.rules.first
-    assert Pundit.authorize(@user, profile.rules.first, :index?)
-    assert Pundit.authorize(@user, profile.rules.first, :show?)
+    assert_equal Rule.all, Pundit.policy_scope(@user, Rule)
+    assert Pundit.authorize(@user, benchmark.rules, :index?)
+    assert Pundit.authorize(@user, benchmark.rules.sample, :show?)
   end
 end