**Before:**
```
EXPLAIN ANALYZE SELECT COUNT(rules.id) FROM rules WHERE rules.id IN (SELECT rules.id FROM rules INNER JOIN profile_rules ON profile_rules.rule_id = rules.id INNER JOIN profiles ON profiles.id = profile_rules.profile_id WHERE (profiles.account_id = 'XY' OR profiles.id IN (SELECT profiles.id FROM profiles WHERE profiles.parent_profile_id IS NULL)))

QUERY PLAN
Finalize Aggregate  (cost=51529.28..51529.29 rows=1 width=8) (actual time=251.133..255.783 rows=1 loops=1)
  ->  Gather  (cost=51529.07..51529.28 rows=2 width=8) (actual time=251.125..255.777 rows=3 loops=1)
        Workers Planned: 2
        Workers Launched: 2
        ->  Partial Aggregate  (cost=50529.07..50529.08 rows=1 width=8) (actual time=244.843..244.849 rows=1 loops=3)
              ->  Parallel Hash Semi Join  (cost=36561.30..50519.81 rows=3702 width=16) (actual time=235.051..244.439 rows=4561 loops=3)
                    Hash Cond: (rules.id = profile_rules.rule_id)
                    ->  Parallel Index Only Scan using rules_pkey on rules  (cost=0.41..12221.57 rows=17061 width=16) (actual time=0.019..7.591 rows=13649 loops=3)
                          Heap Fetches: 11297
                    ->  Parallel Hash  (cost=32457.66..32457.66 rows=212178 width=32) (actual time=218.312..218.316 rows=19158 loops=3)
                          Buckets: 65536  Batches: 16  Memory Usage: 800kB
                          ->  Parallel Hash Join  (cost=13549.26..32457.66 rows=212178 width=32) (actual time=20.604..207.200 rows=19158 loops=3)
                                Hash Cond: (profile_rules.rule_id = rules_1.id)
                                ->  Hash Join  (cost=1114.44..19465.81 rows=212178 width=16) (actual time=7.414..184.843 rows=19158 loops=3)
                                      Hash Cond: (profile_rules.profile_id = profiles.id)
                                      ->  Parallel Seq Scan on profile_rules  (cost=0.00..17236.79 rows=424279 width=32) (actual time=0.006..76.314 rows=339390 loops=3)
                                      ->  Hash  (cost=1079.41..1079.41 rows=2802 width=16) (actual time=6.144..6.146 rows=515 loops=3)
                                            Buckets: 4096  Batches: 1  Memory Usage: 57kB
                                            ->  Seq Scan on profiles  (cost=517.37..1079.41 rows=2802 width=16) (actual time=0.538..6.019 rows=515 loops=3)
                                                  Filter: ((account_id = 'XY'::uuid) OR (hashed SubPlan 1))
                                                  Rows Removed by Filter: 5085
                                                  SubPlan 1
                                                    ->  Bitmap Heap Scan on profiles profiles_1  (cost=12.27..516.08 rows=515 width=16) (actual time=0.059..0.359 rows=515 loops=3)
                                                          Recheck Cond: (parent_profile_id IS NULL)
                                                          Heap Blocks: exact=189
                                                          ->  Bitmap Index Scan on index_profiles_on_parent_profile_id  (cost=0.00..12.15 rows=515 width=0) (actual time=0.037..0.038 rows=515 loops=3)
                                                                Index Cond: (parent_profile_id IS NULL)
                                ->  Parallel Hash  (cost=12221.57..12221.57 rows=17061 width=16) (actual time=12.502..12.502 rows=13649 loops=3)
                                      Buckets: 65536  Batches: 1  Memory Usage: 2496kB
                                      ->  Parallel Index Only Scan using rules_pkey on rules rules_1  (cost=0.41..12221.57 rows=17061 width=16) (actual time=0.024..6.898 rows=13649 loops=3)
                                            Heap Fetches: 11297
Planning Time: 1.469 ms
Execution Time: 255.941 ms

```


**After:**

```
EXPLAIN ANALYZE SELECT COUNT(id) FROM rules

QUERY PLAN
Aggregate  (cost=12570.82..12570.83 rows=1 width=8) (actual time=53.067..53.068 rows=1 loops=1)
  ->  Index Only Scan using rules_pkey on rules  (cost=0.41..12468.45 rows=40946 width=16) (actual time=0.040..50.935 rows=40946 loops=1)
        Heap Fetches: 11061
Planning Time: 3.659 ms
Execution Time: 53.142 ms

```


## Secure Coding Practices Checklist GitHub Link
- https://github.com/RedHatInsights/secure-coding-checklist

## Secure Coding Checklist
- [ ] Input Validation
- [ ] Output Encoding
- [ ] Authentication and Password Management
- [ ] Session Management
- [ ] Access Control
- [ ] Cryptographic Practices
- [ ] Error Handling and Logging
- [ ] Data Protection
- [ ] Communication Security
- [ ] System Configuration
- [ ] Database Security
- [ ] File Management
- [ ] Memory Management
- [x] General Coding Practices