@skateman 2023-01-13
I'm still dealing with an N+1 here...